Moved grid area and clean up
