+++ libs/dh/core/shell/src/lib/dh-core-shell.routes.ts
@@ -84,7 +84,7 @@ export const dhCoreShellRoutes: Routes = [
       },
       {
         path: getPath<BasePaths>('grid-areas'),
-        loadComponent: () => import('@energinet-datahub/dh/market-participant/grid-areas'),
+        loadComponent: () => import('@energinet-datahub/dh/grid-areas'),
         data: {
           titleTranslationKey: 'marketParticipant.gridAreas.topBarTitle',
         },
