+++ libs/dh/grid-areas/jest.config.ts
@@ -18,10 +18,10 @@
 //#endregion
 /* eslint-disable */
 export default {
-  displayName: 'dh-market-participant-grid-areas',
+  displayName: 'dh-grid-areas',
   preset: '../../../../jest.preset.js',
-  setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-  coverageDirectory: '../../../../../coverage/libs/dh/market-participant/grid-areas',
+  setupFilesAfterEnv: ['<rootDir>/tests/test-setup.ts'],
+  coverageDirectory: '../../../../../coverage/libs/dh/grid-areas',
   transform: {
     '^.+\\.(ts|mjs|js|html)$': [
       'jest-preset-angular',
