+++ libs/dh/grid-areas/src/components/details.component.ts
