+++ libs/dh/grid-areas/src/graphql/query/get-grid-area-details.graphql
