+++ libs/dh/grid-areas/src/graphql/query/get-grid-area-overview.graphql
