+++ libs/dh/grid-areas/src/types.ts
