+++ libs/dh/grid-areas/tsconfig.json
@@ -19,7 +19,7 @@
       "path": "./tsconfig.spec.json"
     }
   ],
-  "extends": "../../../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "angularCompilerOptions": {
     "enableI18nLegacyMessageIdFormat": false,
     "strictInjectionParameters": true,
