--- libs/dh/market-participant/grid-areas/src/graphql/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["../../../../../../.eslintrc.json"],
-  "ignorePatterns": ["!**/*"]
-}
